+++ tensorpc/flow/components/flowplus/__init__.py
@@ -1,5 +1,7 @@
 from . import customnode, defaultnodes
-from .compute import (ComputeFlow, ComputeNode, register_compute_node,
-                      schedule_next, schedule_node, NodeConfig, WrapperConfig,
-                      DontSchedule)
+from .compute import (ComputeFlow, ComputeNode, DontSchedule, NodeConfig,
+                      WrapperConfig, register_compute_node, schedule_next,
+                      schedule_node)
 from .defaultnodes import ResizeableNodeBase
+from .processutil import ProcessPoolExecutor as NodeProcessPoolExecutor
+from .processutil import run_in_node_executor