+++ tensorpc/flow/components/plus/arraycommon.py
@@ -10,6 +10,9 @@ def can_cast_to_np_array(obj: Any):
         return True
     elif get_qualname_of_type(type(obj)) == CommonQualNames.TorchTensor:
         return True
+    elif get_qualname_of_type(type(obj)) == CommonQualNames.TorchParameter:
+        return True
+
     elif get_qualname_of_type(type(obj)) == CommonQualNames.TVTensor:
         return True
     return False
@@ -20,6 +23,8 @@ def try_cast_to_np_array(obj: Any) -> Optional[np.ndarray]:
         return obj
     elif get_qualname_of_type(type(obj)) == CommonQualNames.TorchTensor:
         return obj.detach().cpu().numpy()
+    elif get_qualname_of_type(type(obj)) == CommonQualNames.TorchParameter:
+        return obj.data.detach().cpu().numpy()
     elif get_qualname_of_type(type(obj)) == CommonQualNames.TVTensor:
         return obj.cpu().numpy()
     return None