+++ tensorpc/utils/typeutils.py
@@ -6,12 +6,12 @@ P = ParamSpec('P')
 
 
 def take_annotation_from(
-    this: Callable[P, Optional[T]]
-) -> Callable[[Callable], Callable[P, Optional[T]]]:
+    this: Callable[P, T]
+) -> Callable[[Callable], Callable[P, T]]:
 
-    def decorator(real_function: Callable) -> Callable[P, Optional[T]]:
+    def decorator(real_function: Callable) -> Callable[P, T]:
 
-        def new_function(*args: P.args, **kwargs: P.kwargs) -> Optional[T]:
+        def new_function(*args: P.args, **kwargs: P.kwargs) -> T:
             return real_function(*args, **kwargs)
 
         return new_function